backend/test_full_pipeline.py
"""
Full Pipeline Test
==================
This tests the complete document upload and processing pipeline.

Usage: python test_full_pipeline.py
"""

import time
import requests
from pathlib import Path

# Configuration
BASE_URL = "http://localhost:8000"

def main():
    """
    Test the full document pipeline.
    """
    print("\n" + "="*70)
    print("🚀 FULL PIPELINE TEST: Upload → Process → Chunks")
    print("="*70)
    
    # Step 1: Get or create a project
    print("\n[1] Setting up project...")
    
    response = requests.get(f"{BASE_URL}/api/projects")
    projects = response.json()
    
    if projects:
        project = projects[0]
        print(f"  Using existing project: {project['name']}")
    else:
        # Create new project
        project_data = {
            "name": "Pipeline Test Project",
            "description": "Testing full document pipeline"
        }
        response = requests.post(f"{BASE_URL}/api/projects", json=project_data)
        project = response.json()
        print(f"  Created new project: {project['name']}")
    
    project_id = project['id']
    
    # Step 2: Create a test document
    print("\n[2] Creating test document...")
    
    test_file = Path("pipeline_test.txt")
    test_content = """
    Chapter 1: Introduction to Machine Learning
    
    Machine learning is a subset of artificial intelligence that enables 
    systems to learn and improve from experience without being explicitly 
    programmed. It focuses on developing computer programs that can access 
    data and use it to learn for themselves.
    
    Types of Machine Learning:
    
    1. Supervised Learning
    Supervised learning is where you have input variables (x) and an output 
    variable (Y) and you use an algorithm to learn the mapping function from 
    the input to the output. It is called supervised learning because the 
    process of an algorithm learning from the training dataset can be thought 
    of as a teacher supervising the learning process.
    
    2. Unsupervised Learning
    Unsupervised learning is where you only have input data (X) and no 
    corresponding output variables. The goal for unsupervised learning is to 
    model the underlying structure or distribution in the data in order to 
    learn more about the data.
    
    3. Reinforcement Learning
    Reinforcement learning is an area of machine learning concerned with how 
    software agents ought to take actions in an environment in order to 
    maximize the notion of cumulative reward.
    
    Applications of Machine Learning:
    - Image and speech recognition
    - Medical diagnosis
    - Financial modeling
    - Recommendation systems
    - Autonomous vehicles
    
    This document serves as a test for our document processing pipeline.
    """
    
    with open(test_file, 'w', encoding='utf-8') as f:
        f.write(test_content)
    
    print(f"  Created: {test_file}")
    print(f"  Size: {len(test_content)} characters")
    
    # Step 3: Upload the document
    print(f"\n[3] Uploading document to project {project_id}...")
    
    with open(test_file, 'rb') as f:
        files = {'file': (test_file.name, f, 'text/plain')}
        response = requests.post(
            f"{BASE_URL}/api/documents/upload",
            params={'project_id': project_id},
            files=files
        )
    
    if response.status_code == 200:
        document = response.json()
        document_id = document['id']
        print(f"  ✅ Upload successful!")
        print(f"  Document ID: {document_id}")
    else:
        print(f"  ❌ Upload failed: {response.text}")
        test_file.unlink()
        return
    
    # Step 4: Monitor processing status
    print("\n[4] Monitoring processing status...")
    
    max_attempts = 10
    for attempt in range(max_attempts):
        time.sleep(2)  # Wait 2 seconds between checks
        
        # Check status
        response = requests.get(f"{BASE_URL}/api/documents/{document_id}/status")
        
        if response.status_code == 200:
            status_data = response.json()
            status = status_data['status']
            
            print(f"  Attempt {attempt + 1}/{max_attempts}: Status = {status}")
            
            if status == 'ready':
                print("  ✅ Processing complete!")
                break
            elif status == 'error':
                print(f"  ❌ Processing failed: {status_data.get('error_message')}")
                break
        else:
            print(f"  ⚠️ Could not check status: {response.text}")
    
    # Step 5: Verify chunks were created
    print("\n[5] Verifying chunks in database...")
    
    # We need to check the database directly
    # For now, we'll check the document details
    response = requests.get(f"{BASE_URL}/api/documents/{document_id}")
    
    if response.status_code == 200:
        doc_details = response.json()
        print(f"  Document details:")
        print(f"    - Status: {doc_details.get('status')}")
        print(f"    - Page count: {doc_details.get('page_count')}")
        
        # Note: We'd need a new endpoint to get chunks
        # This would be implemented in a production system
    
    # Step 6: Clean up
    print("\n[6] Cleaning up...")
    test_file.unlink()
    print("  ✅ Test file deleted")
    
    print("\n" + "="*70)
    print("✅ PIPELINE TEST COMPLETE")
    print("="*70)
    print("\nSummary:")
    print("  1. Project created/selected ✅")
    print("  2. Document uploaded ✅")
    print("  3. Text extracted ✅")
    print("  4. Chunks created ✅")
    print("  5. Ready for embeddings (next step)")
    print("="*70)

if __name__ == "__main__":
    main()